Implement Windows and Linux monitor and window backends
- Added Win32 monitor backend with accurate monitor enumeration and properties.
- Refactored Win32 window backend to manage window creation, destruction, and events.
- Introduced Wayland monitor and window backends for Linux, supporting monitor enumeration and window management.
- Created X11 monitor and window backends for Linux, enabling compatibility with X11 systems.
- Enhanced timer resolution on Windows for improved sleep/timer accuracy.
- Implemented event staging and flushing mechanisms in the Win32 window backend to handle modal loops.
- Added necessary headers and structures for Wayland and X11 backends.
