+++ src/core/utility/thread.cpp
@@ -52,7 +52,13 @@ void PreciseSleepNs(u64 nanoseconds) noexcept
     return;
   }
 
-  const u64 spinThresholdNs = 100000;
+  // On Windows the default timer resolution is ~15.6ms, so short sleeps
+  // overshoot massively.  Use a larger spin threshold to compensate.
+#if defined(_WIN32)
+  const u64 spinThresholdNs = 2000000;  // 2ms — stay within one timer tick
+#else
+  const u64 spinThresholdNs = 100000;   // 100us — Linux/macOS are fine
+#endif
 
   if (nanoseconds > spinThresholdNs)
   {
