+++ src/platform/platform_module.cpp
@@ -3,6 +3,15 @@
 #include "gecko/core/scope.h"
 #include "gecko/core/services/log.h"
 
+#if defined(_WIN32)
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
+#include <Windows.h>
+#include <timeapi.h>
+#pragma comment(lib, "Winmm.lib")
+#endif
+
 namespace gecko::platform {
 
 static PlatformModule s_PlatformModule;
@@ -10,13 +19,23 @@ static PlatformModule s_PlatformModule;
 bool PlatformModule::Startup(::gecko::IModuleRegistry& /*modules*/) noexcept
 {
   GECKO_FUNC(labels::Platform);
-  // Minimal placeholder extension point.
+
+#if defined(_WIN32)
+  // Request 1ms timer resolution so sleep/timer functions are accurate.
+  // Without this, Sleep(16) rounds up to ~31ms on default 15.6ms ticks.
+  ::timeBeginPeriod(1);
+#endif
+
   return true;
 }
 
 void PlatformModule::Shutdown(::gecko::IModuleRegistry& /*modules*/) noexcept
 {
   GECKO_FUNC(labels::Platform);
+
+#if defined(_WIN32)
+  ::timeEndPeriod(1);
+#endif
 }
 
 ::gecko::ModuleRegistration InstallPlatformModule(
