+++ src/platform/win32/win32_windows_backend.cpp
@@ -1,4 +1,4 @@
-#include "gecko/platform/platform_config.h"
+#include "win32_windows_backend.h"
 
 #if defined(_WIN32)
 
@@ -6,30 +6,12 @@
 #include "../private/platform_utils.h"
 #include "gecko/core/ptr.h"
 #include "gecko/core/scope.h"
-#include "gecko/core/services/events.h"
 #include "gecko/core/services/log.h"
 #include "gecko/platform/input_codes.h"
-#include "gecko/platform/platform_events.h"
-#include "gecko/platform/window.h"
-#include "gecko/platform/windows_interface.h"
 
-#ifndef WIN32_LEAN_AND_MEAN
-#define WIN32_LEAN_AND_MEAN
-#endif
-#ifndef NOMINMAX
-#define NOMINMAX
-#endif
 #include <cstring>
 #include <shellscalingapi.h>
-#include <string>
-#include <unordered_map>
-#include <vector>
-#include <Windows.h>
 #include <windowsx.h>
-
-// Windows.h defines CreateWindow as a macro (CreateWindowA/W), which collides
-// with our IWindowsBackend::CreateWindow virtual method.
-#undef CreateWindow
 
 #pragma comment(lib, "Shcore.lib")
 #pragma comment(lib, "User32.lib")
@@ -135,128 +117,6 @@ MouseButton WmButtonToMouseButton(UINT msg) noexcept
 constexpr wchar_t kWndClassName[] = L"GeckoWindowClass";
 
 }  // namespace
-
-// ╔══════════════════════════════════════════════════════════════════════╗
-// ║  Win32WindowsBackend                                                ║
-// ╚══════════════════════════════════════════════════════════════════════╝
-
-class Win32WindowsBackend final : public IWindowsBackend
-{
-public:
-  Win32WindowsBackend() noexcept;
-  ~Win32WindowsBackend() noexcept override;
-
-  WindowHandle CreateWindow(const WindowDesc& desc) noexcept override;
-  void DestroyWindow(WindowHandle window) noexcept override;
-  bool IsWindowAlive(WindowHandle window) const noexcept override;
-  bool RequestClose(WindowHandle window) noexcept override;
-
-  Extent2D GetClientSize(WindowHandle window) const noexcept override;
-  void SetClientSize(WindowHandle window, Extent2D size) noexcept override;
-  void SetTitle(WindowHandle window, const char* title) noexcept override;
-  const char* GetTitle(WindowHandle window) const noexcept override;
-  void SetPosition(WindowHandle window, math::Int2 pos) noexcept override;
-  math::Int2 GetPosition(WindowHandle window) const noexcept override;
-  DpiInfo GetDpi(WindowHandle window) const noexcept override;
-  NativeWindowHandle GetNativeWindowHandle(
-      WindowHandle window) const noexcept override;
-
-  void SetWindowState(WindowHandle window,
-                      platform::WindowState state) noexcept override;
-  platform::WindowState GetWindowState(
-      WindowHandle window) const noexcept override;
-  void SetDecorated(WindowHandle window, bool decorated) noexcept override;
-  bool IsDecorated(WindowHandle window) const noexcept override;
-  void RequestFocus(WindowHandle window) noexcept override;
-
-  void SetResizable(WindowHandle window, bool resizable) noexcept override;
-  bool IsResizable(WindowHandle window) const noexcept override;
-  void SetWindowMode(WindowHandle window, WindowMode mode) noexcept override;
-  WindowMode GetWindowMode(WindowHandle window) const noexcept override;
-  void SetWindowButtons(WindowHandle window,
-                        WindowButtons buttons) noexcept override;
-  WindowButtons GetWindowButtons(
-      WindowHandle window) const noexcept override;
-  void SetMinSize(WindowHandle window, Extent2D size) noexcept override;
-  void SetMaxSize(WindowHandle window, Extent2D size) noexcept override;
-  void SetAlwaysOnTop(WindowHandle window, bool topmost) noexcept override;
-  bool IsAlwaysOnTop(WindowHandle window) const noexcept override;
-
-  void SetCursorMode(WindowHandle window, CursorMode mode) noexcept override;
-  CursorMode GetCursorMode(WindowHandle window) const noexcept override;
-
-  void PumpEvents(const gecko::EventEmitter& emitter) noexcept override;
-
-private:
-  struct Win32WindowEntry
-  {
-    WindowDesc Desc {};
-    HWND Hwnd {nullptr};
-    Extent2D ClientSize {};
-    math::Int2 Position {};
-    platform::WindowState State {platform::WindowState::Normal};
-    CursorMode Cursor {CursorMode::Normal};
-    WindowMode Mode {WindowMode::Windowed};
-    WindowButtons Buttons {WindowButtons::All};
-    Extent2D MinSize {0, 0};
-    Extent2D MaxSize {0, 0};
-    bool Decorated {true};
-    bool Resizable {true};
-    bool AlwaysOnTop {false};
-    bool Alive {true};
-    std::string TitleStorage;
-    // Saved style/position for fullscreen restoration.
-    DWORD SavedStyle {0};
-    DWORD SavedExStyle {0};
-    RECT SavedRect {};
-  };
-
-  struct StagedEvent
-  {
-    gecko::EventCode Code {};
-    union PayloadUnion
-    {
-      events::WindowClosedPayload Closed;
-      events::WindowCloseRequestedPayload CloseRequested;
-      events::WindowResizedPayload Resized;
-      events::WindowDpiChangedPayload DpiChanged;
-      events::WindowKeyPayload Key;
-      events::WindowCharPayload Char;
-      events::WindowMouseMovePayload MouseMove;
-      events::WindowMouseButtonPayload MouseButton;
-      events::WindowMouseWheelPayload MouseWheel;
-      events::WindowFocusChangedPayload FocusChanged;
-      events::WindowMovedPayload Moved;
-      events::WindowStateChangedPayload StateChanged;
-      PayloadUnion() noexcept
-      {
-        std::memset(this, 0, sizeof(PayloadUnion));
-      }
-    } Data;
-    u32 PayloadSize {0};
-  };
-
-  Win32WindowEntry* FindEntry(WindowHandle window) noexcept;
-  const Win32WindowEntry* FindEntry(WindowHandle window) const noexcept;
-  Win32WindowEntry* FindByHwnd(HWND hwnd) noexcept;
-  WindowHandle HandleFromHwnd(HWND hwnd) const noexcept;
-
-  void StageEvent(const StagedEvent& ev) noexcept;
-  static LRESULT CALLBACK WndProc(HWND hwnd, UINT msg, WPARAM wParam,
-                                  LPARAM lParam);
-
-  void ApplyDecorations(HWND hwnd, bool decorated, bool resizable) noexcept;
-  DWORD MakeStyle(const WindowDesc& desc) const noexcept;
-
-  u64 m_NextId {0};
-  ATOM m_WndClass {0};
-  std::unordered_map<u64, Win32WindowEntry> m_Windows;
-  std::vector<StagedEvent> m_Staged;
-
-  // We need the backend pointer inside the static WndProc.
-  // We store a global instance pointer because all windows share one backend.
-  static Win32WindowsBackend* s_Instance;
-};
 
 Win32WindowsBackend* Win32WindowsBackend::s_Instance = nullptr;
 
@@ -868,6 +728,8 @@ CursorMode Win32WindowsBackend::GetCursorMode(
 void Win32WindowsBackend::PumpEvents(
     const gecko::EventEmitter& emitter) noexcept
 {
+  m_CurrentEmitter = &emitter;
+
   MSG msg;
   while (::PeekMessageW(&msg, nullptr, 0, 0, PM_REMOVE))
   {
@@ -875,10 +737,8 @@ void Win32WindowsBackend::PumpEvents(
     ::DispatchMessageW(&msg);
   }
 
-  for (const auto& ev : m_Staged)
-    gecko::SendEvent(emitter, ev.Code,
-                     gecko::EventView {&ev.Data, ev.PayloadSize});
-  m_Staged.clear();
+  FlushStagedEvents();
+  m_CurrentEmitter = nullptr;
 }
 
 // ── WndProc ────────────────────────────────────────────────────────────
@@ -1131,6 +991,29 @@ LRESULT CALLBACK Win32WindowsBackend::WndProc(HWND hwnd, UINT msg,
     break;
   }
 
+  case WM_ENTERSIZEMOVE: {
+    // Windows enters a modal loop during drag/resize that blocks our
+    // PumpEvents.  Start a fast timer so we can keep flushing staged
+    // events (resize, move, etc.) while the modal loop is running.
+    ::SetTimer(hwnd, kModalTimerId, kModalTimerIntervalMs, nullptr);
+    break;
+  }
+
+  case WM_EXITSIZEMOVE: {
+    ::KillTimer(hwnd, kModalTimerId);
+    break;
+  }
+
+  case WM_TIMER: {
+    if (wParam == kModalTimerId)
+    {
+      self->FlushStagedEvents();
+      if (self->m_ModalFrameCallback)
+        self->m_ModalFrameCallback(self->m_ModalFrameUserData);
+    }
+    break;
+  }
+
   default:
     break;
   }
@@ -1177,6 +1060,24 @@ WindowHandle Win32WindowsBackend::HandleFromHwnd(HWND hwnd) const noexcept
 void Win32WindowsBackend::StageEvent(const StagedEvent& ev) noexcept
 {
   m_Staged.push_back(ev);
+}
+
+void Win32WindowsBackend::FlushStagedEvents() noexcept
+{
+  if (!m_CurrentEmitter || m_Staged.empty())
+    return;
+
+  for (const auto& ev : m_Staged)
+    gecko::SendEvent(*m_CurrentEmitter, ev.Code,
+                     gecko::EventView {&ev.Data, ev.PayloadSize});
+  m_Staged.clear();
+}
+
+void Win32WindowsBackend::SetModalFrameCallback(ModalFrameFn callback,
+                                                void* userData) noexcept
+{
+  m_ModalFrameCallback = callback;
+  m_ModalFrameUserData = userData;
 }
 
 // ── Factory ────────────────────────────────────────────────────────────
